+++ lambda/api/subscribe_folder/function.py
@@ -2,6 +2,10 @@
 import json
 from helper import subscribe_folder # pylint: disable=import-error
 
+from helper import maintenance_guard # pylint: disable=import-error
+
+
+@maintenance_guard
 def handler(event, _context):
     '''Marks the specified folder as subscribed'''
     body = json.loads(event['body'])